+++ CompanyManagementSystemRework/cli/EmployeeModule.py
@@ -1,13 +1,38 @@
 import pandas as pd
+import UtilityModule
 
 class EmployeeModule:
     def __init__(self):
         self.file_name = "Datas/Employee_Dataset.csv"
-        self.load_data()
+        self.employee_data = pd.read_csv(self.file_name)
 
     def createEmployee(self):
-        # Add Employee
-        pass 
+        '''
+        Creates a new employee through user input and adds new employee data to the current list
+        '''
+        first_name = input("Enter First Name:") 
+        last_name =  input("Enter Last Name")
+        position =  input("Enter Position:")
+        gender = input("Enter Gender:")
+        salary = input("Enter Salary:")
+
+        Employee_Data = {
+            "First Name" : first_name,
+            "Last Name" : last_name,
+            "Positions" : position.upper(),
+            "Gender" : gender.capitalize(),
+            "Salary" : int(salary)
+        }
+
+        self.employee_data = self.employee_data.append(self.employee_data, ignore_index=True)
+
+        if (input("Show new employee list? [Y for yes]:").upper() == "Y"):
+            self.showEmployee(full_length=True)
+        
+        self.employee_data.to_csv(self.file_name)
+
+        
+        
 
     def editEmployee(self):
         # Search for Employee First
@@ -98,77 +123,20 @@ class EmployeeModule:
         return search_df
     
     # DONE NC
-    def showEmployee(self, id = 0, singular = False):
+    def showEmployee(self, singular = False, full_length = False):
         # Just calls the print dataframe function
-        FULL_LENGTH = False
-        if FULL_LENGTH:
+        if full_length:
             len_print = len(self.employee_data[self.employee_data.columns.tolist()[0]])
         else:
             len_print = 100
 
-        self.print_dataframe(self.employee_data, len_print, 3)
+        UtilityModule.print_dataframe(self.employee_data, len_print)
 
-    # UTILITY FUNCTIONS 
-    def print_dataframe(self, df, UPPER_LIMIT = 100, PADDING = 3):
-        '''
-        Printing the given dataframe, can be any of the following:
-        - Actual dataframe of the csv file
-        - Cut dataframe (defined by UPPER LIMIT VARIABLE)
-        - Modified dataframe from search command
-        '''
-
-        # Find the lengths to be used in right-justify
-        maxes = self.find_length_per_column(df)
-        
-        # Print a blank element to space the columns accordingly
-        print("".ljust(len(str(UPPER_LIMIT))), end = "")
-        for index, element in enumerate(df.columns.tolist()):
-            print(element.rjust(maxes[index] + PADDING), end = "  ")
-        print()
-
-        # Printing of each row
-        for row in range(0, UPPER_LIMIT):
-
-            # If printing the search dataframe, there would be blank indexes
-            # This would be for skipping those
-            if row not in df.index:
-                continue 
-
-            print(str(row).ljust(len(str(UPPER_LIMIT))), end = "")
-            for index, element in enumerate(df.columns.tolist()):
-                length_spacing = maxes[index] + PADDING
-                print(str(df[element][row]).rjust(length_spacing),end ="  ")
-
-            print()
-
-    def find_length_per_column(self, df):
-        '''
-        Finds the length of the texts in each column which will be used as length for the right-justify (rjust)
-        Purpose: 
-        Dataframe printing by line since pandas does not have a print by line without showing
-        the column above it
-        '''
-        columns = df.columns.tolist()
-        lengths = []
-
-        for col in columns:
-            a = [len(str(i)) for i in df[col].tolist()]
-            lengths.append(max(a))
-
-        return lengths
-
-    def save_data(self):
-        self.employee_data.to_csv(self.file_name)
     
-    def load_data(self):
-        self.employee_data = pd.read_csv(self.file_name)
-        try:
-            self.employee_data.drop("#", axis=1, inplace=True)
-        except:
-            pass
         
          
 
 if __name__ == "__main__":
     EM = EmployeeModule()
-    EM.searchEmployee()
+    EM.createEmployee()
+    #EM.showEmployee()
